Correctly detect sprites that are on the right screen border
Overflow is a hell of a drug...

+++ src/ppu.rs
@@ -348,7 +348,7 @@ impl SpriteStruct {
 
     // Quick test to see whether the given point is in the bounding box of this sprite.
     fn in_bounding_box(&self, ppu: &Ppu, x: u8, y: u8) -> bool {
-        x >= self.x && x < self.x + 8 && self.on_scanline(ppu, y)
+        x >= self.x && (x as u16) < (self.x as u16 + 8) && self.on_scanline(ppu, y)
     }
 }
